+++ TP3_areaCalculator/src/main/java/Elipse.java
@@ -1,31 +1,31 @@
 public class Elipse implements AreaShape {
-    private float x_radius;
-    private float y_radius;
+    private double x_radius;
+    private double y_radius;
 
-    Elipse(float x_radius, float y_radius){
+    Elipse(double x_radius, double y_radius){
         this.x_radius=x_radius;
         this.y_radius=y_radius;
     }
 
-    public float getX_radius() {
+    public double getX_radius() {
         return x_radius;
     }
 
-    public void setX_radius(float x_radius) {
+    public void setX_radius(double x_radius) {
         this.x_radius = x_radius;
     }
 
-    public float getY_radius() {
+    public double getY_radius() {
         return y_radius;
     }
 
-    public void setY_radius(float y_radius) {
+    public void setY_radius(double y_radius) {
         this.y_radius = y_radius;
     }
 
     @Override
-    public float getArea() {
-        return (float) (Math.PI * x_radius * y_radius);
+    public double getArea() {
+        return (Math.PI * x_radius * y_radius);
     }
 
     @Override
